pyproject.toml
[project]
name = "slack-excel-bot"
version = "0.1.0"
description = "Slack chatbot that uses OpenAI tool calling to generate Excel templates."
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.10.0",
  "fastapi>=0.111.0",
  "httpx>=0.27.0",
  "openai>=1.75.0",
  "openpyxl>=3.1.2",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
  "slack_sdk>=3.33.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["slack_excel_bot*"]

[tool.setuptools.package-data]
slack_excel_bot = [
  "templates/**/*.json",
  "templates/**/*.yaml",
  "templates/**/*.xlsx",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
